# apps/api/contract/openapi.yaml
openapi: 3.1.0
info:
  title: LearnR API
  version: 0.1.0
...
paths:
...
  /children/{id}/record:
    get:
      operationId: readChildRecord
      parameters:
        - schema:
            default: maths
            type: string
          in: query
          name: subject
          required: false
        - schema:
            default: 3
            type: integer
            minimum: 1
            maximum: 50
          in: query
          name: perTopic
          required: false
        - schema:
            default: 2505600000
            type: integer
            minimum: 1
            maximum: 2505600000
          in: query
          name: windowMs
          required: false
        - schema:
            default: 'false'
            type: string
            enum:
              - 'true'
              - 'false'
          in: query
          name: speedRuns
          required: false
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  observations:
                    type: array
                    items:
                      type: object
                      properties:
                        topic:
                          type: string
                        level:
                          type: string
                          enum:
                            - '1'
                            - '2'
                            - '3'
                            - '4'
                            - '5'
                            - '6'
                            - K
                        correct:
                          type: boolean
                        templateId:
                          type: string
                        timeTakenMs:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        answeredAt:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        offsetMinutes:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                      required:
                        - topic
                        - level
                        - correct
                        - timeTakenMs
                        - answeredAt
                      additionalProperties: false
                  sittings:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        startedAt:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        level:
                          type: string
                          enum:
                            - '1'
                            - '2'
                            - '3'
                            - '4'
                            - '5'
                            - '6'
                            - K
                        attempts:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        correct:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        timeMs:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                      required:
                        - id
                        - startedAt
                        - level
                        - attempts
                        - correct
                        - timeMs
                      additionalProperties: false
                  answers:
                    anyOf:
                      - type: array
                        items:
                          type: object
                          properties:
                            topic:
                              type: string
                            level:
                              type: string
                              enum:
                                - '1'
                                - '2'
                                - '3'
                                - '4'
                                - '5'
                                - '6'
                                - K
                            prompt:
                              type: string
                            expected:
                              type: string
                            response:
                              type: string
                            correct:
                              type: boolean
                            answeredAt:
                              type: integer
                              minimum: -9007199254740991
                              maximum: 9007199254740991
                            figure:
                              type: object
                              properties:
                                width:
                                  type: number
                                height:
                                  type: number
                                marks:
                                  readOnly: true
                                  maxItems: 200
                                  type: array
                                  items:
                                    oneOf:
                                      - type: object
                                        properties:
                                          kind:
                                            type: string
                                            enum:
                                              - path
                                          points:
                                            readOnly: true
                                            type: array
                                            items:
                                              readOnly: true
                                              type: array
                                              prefixItems:
                                                - type: number
                                                - type: number
                                          closed:
                                            type: boolean
                                          fill:
                                            type: boolean
                                          dashed:
                                            type: boolean
                                        required:
                                          - kind
                                          - points
                                          - closed
                                          - fill
                                          - dashed
                                        additionalProperties: false
                                      - type: object
                                        properties:
                                          kind:
                                            type: string
                                            enum:
                                              - arc
                                          at:
                                            readOnly: true
                                            type: array
                                            prefixItems:
                                              - type: number
                                              - type: number
                                          radius:
                                            type: number
                                          from:
                                            type: number
                                          to:
                                            type: number
                                        required:
                                          - kind
                                          - at
                                          - radius
                                          - from
                                          - to
                                        additionalProperties: false
                                      - type: object
                                        properties:
                                          kind:
                                            type: string
                                            enum:
                                              - dot
                                          at:
                                            readOnly: true
                                            type: array
                                            prefixItems:
                                              - type: number
                                              - type: number
                                        required:
                                          - kind
                                          - at
                                        additionalProperties: false
                                      - type: object
                                        properties:
                                          kind:
                                            type: string
                                            enum:
                                              - label
                                          at:
                                            readOnly: true
                                            type: array
                                            prefixItems:
                                              - type: number
                                              - type: number
                                          text:
                                            type: string
                                        required:
                                          - kind
                                          - at
                                          - text
                                        additionalProperties: false
                              required:
                                - width
                                - height
                                - marks
                              additionalProperties: false
                          required:
                            - topic
                            - level
                            - prompt
                            - expected
                            - response
                            - correct
                            - answeredAt
                          additionalProperties: false
                      - type: 'null'
                  recentAnswers:
                    anyOf:
                      - type: array
                        items:
                          type: object
                          properties:
                            answeredAt:
                              type: integer
                              minimum: -9007199254740991
                              maximum: 9007199254740991
                            timeTakenMs:
                              type: integer
                              minimum: -9007199254740991
                              maximum: 9007199254740991
                          required:
                            - answeredAt
                            - timeTakenMs
                          additionalProperties: false
                      - type: 'null'
                  speedRuns:
                    anyOf:
                      - type: array
                        items:
                          type: object
                          properties:
                            mode:
                              type: string
                            correct:
                              type: integer
                              minimum: -9007199254740991
                              maximum: 9007199254740991
                            playedAt:
                              type: string
                              format: date-time
                          required:
                            - mode
                            - correct
                            - playedAt
                          additionalProperties: false
                      - type: 'null'
                required:
                  - observations
                  - sittings
                  - answers
                  - recentAnswers
                  - speedRuns
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
        '503':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /children/{id}/answers:
    get:
      operationId: readAnsweredQuestions
      parameters:
        - schema:
            default: maths
            type: string
          in: query
          name: subject
          required: false
        - schema:
            default: 3
            type: integer
            minimum: 1
            maximum: 50
          in: query
          name: perTopic
          required: false
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    topic:
                      type: string
                    level:
                      type: string
                      enum:
                        - '1'
                        - '2'
                        - '3'
                        - '4'
                        - '5'
                        - '6'
                        - K
                    prompt:
                      type: string
                    expected:
                      type: string
                    response:
                      type: string
                    correct:
                      type: boolean
                    answeredAt:
                      type: integer
                      minimum: -9007199254740991
                      maximum: 9007199254740991
                    figure:
                      type: object
                      properties:
                        width:
                          type: number
                        height:
                          type: number
                        marks:
                          readOnly: true
                          maxItems: 200
                          type: array
                          items:
                            oneOf:
                              - type: object
                                properties:
                                  kind:
                                    type: string
                                    enum:
                                      - path
                                  points:
                                    readOnly: true
                                    type: array
                                    items:
                                      readOnly: true
                                      type: array
                                      prefixItems:
                                        - type: number
                                        - type: number
                                  closed:
                                    type: boolean
                                  fill:
                                    type: boolean
                                  dashed:
                                    type: boolean
                                required:
                                  - kind
                                  - points
                                  - closed
                                  - fill
                                  - dashed
                                additionalProperties: false
                              - type: object
                                properties:
                                  kind:
                                    type: string
                                    enum:
                                      - arc
                                  at:
                                    readOnly: true
                                    type: array
                                    prefixItems:
                                      - type: number
                                      - type: number
                                  radius:
                                    type: number
                                  from:
                                    type: number
                                  to:
                                    type: number
                                required:
                                  - kind
                                  - at
                                  - radius
                                  - from
                                  - to
                                additionalProperties: false
                              - type: object
                                properties:
                                  kind:
                                    type: string
                                    enum:
                                      - dot
                                  at:
                                    readOnly: true
                                    type: array
                                    prefixItems:
                                      - type: number
                                      - type: number
                                required:
                                  - kind
                                  - at
                                additionalProperties: false
                              - type: object
                                properties:
                                  kind:
                                    type: string
                                    enum:
                                      - label
                                  at:
                                    readOnly: true
                                    type: array
                                    prefixItems:
                                      - type: number
                                      - type: number
                                  text:
                                    type: string
                                required:
                                  - kind
                                  - at
                                  - text
                                additionalProperties: false
                      required:
                        - width
                        - height
                        - marks
                      additionalProperties: false
                  required:
                    - topic
                    - level
                    - prompt
                    - expected
                    - response
                    - correct
                    - answeredAt
                  additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
        '503':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
...
  /speed/summaries:
    get:
      operationId: readSpeedSummaries
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    mode:
                      type: string
                    correct:
                      type: integer
                      minimum: -9007199254740991
                      maximum: 9007199254740991
                    playedAt:
                      type: string
                      format: date-time
                  required:
                    - mode
                    - correct
                    - playedAt
                  additionalProperties: false
        '503':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
...